* master:
  Correctly reset pstat->om on dos RESET
  Set endpoint->version for eps sent to client cbs
  Use oc_interface_mask_t not uint8_t
  Added checking error to prevent crash
  Added checking the type of key to prevent crash
  Added to compare option size and data length
  Added checking the invalid type to prevent crash
  Add missing #include for self header
  Add static to internal functions
  Misc. fixes
  Add extern "C" for use in C++ builds
  Add remove observer logic when delete resource.
  Adding brackets at IPv6 address
  Update code to use size_t
  Add Unit tests, fixes, and doc for oc_base64
  oc_pstat: set POST response length to zero
  Address Win64 build warning
  Add casts to macros
  Fix Windows x64 build warnings
  Update Linux Makefile
  Fix crash in parsing heterogeneous array.
  tizenrt: Select LIBM used by tinycbor
  Add check logic in oc_main_shutdown()
  Adding Null check at rep macro
  fix tests/client_get_linux compile fail error.
  UTC for RI
  UTC for core resource
  Drop old observe callback

+++ apps/st_app/linux/st_ref_app.c
@@ -107,7 +107,7 @@ get_resource_handler(st_request_t *request)
 static void
 switch_resource_change(oc_rep_t *rep)
 {
-  int len = 0;
+  size_t len = 0;
   char *m_power = NULL;
   if (oc_rep_get_string(rep, power_prop_key, &m_power, &len)) {
     strncpy(power, m_power, len);
